requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
einops==0.6.0
exceptiongroup==1.1.0
filelock==3.9.0
huggingface-hub==0.11.1
idna==3.4
imutils==0.5.4
iniconfig==2.0.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
Nuitka==1.5.4
numpy==1.24.2
opencv-python==4.7.0.72
ordered-set==4.1.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sympy==1.11.1
timm==0.6.12
tomli==2.0.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
wxPython==4.2.0
XlsxWriter==3.0.8
zstandard==0.20.0
